SelectPickerをiOS、Androidで共通化
