+++ example-app/SantokuApp/src/components/picker/SelectPicker.tsx
@@ -1,11 +1,20 @@
-import {PickerProps} from '@react-native-picker/picker';
-import {log} from 'framework/logging';
 import React from 'react';
-import {Platform, StyleProp, TextStyle, TextInputProps as RNETextInputProps, ViewProps} from 'react-native';
+import {
+  Pressable,
+  StyleProp,
+  StyleSheet,
+  TextInput,
+  TextInputProps as RNETextInputProps,
+  TextStyle,
+  View,
+  ViewProps,
+} from 'react-native';
 
-import {DefaultPickerAccessoryProps} from './DefaultPickerAccessory';
-import {PickerBackdropProps} from './PickerBackdrop';
-import {PickerContainerProps} from './PickerContainer';
+import {DefaultPickerAccessory, DefaultPickerAccessoryProps} from './DefaultPickerAccessory';
+import {PickerBackdrop, PickerBackdropProps} from './PickerBackdrop';
+import {PickerContainer, PickerContainerProps} from './PickerContainer';
+import {SelectPickerItems, SelectPickerItemsProps} from './SelectPickerItems';
+import {useSelectPickerUseCase} from './useSelectPickerUseCase';
 
 export type Item<T> = {
   label: string;
@@ -112,7 +121,7 @@ export type SelectPickerProps<ItemT> = {
    * PickerのProps
    * なお、pickerItemsComponentを指定した場合は使用されません。
    */
-  pickerProps?: Omit<PickerProps<ItemT>, 'selectedValue' | 'onValueChange'>;
+  pickerProps?: Omit<SelectPickerItemsProps<ItemT>, 'items' | 'selectedValue' | 'onValueChange' | 'keyExtractor'>;
   /**
    * PickerBackdropのProps
    *
@@ -135,8 +144,96 @@ export type SelectPickerProps<ItemT> = {
 };
 
 export const SelectPicker = <ItemT extends unknown>(props: SelectPickerProps<ItemT>) => {
-  React.useEffect(() => {
-    log.warn(`SelectPicker is not supported on: ${Platform.OS}`);
-  }, []);
-  return null;
+  const {
+    isVisible,
+    inputValue,
+    handleBackdropPress,
+    pickerBackdropEntering,
+    pickerBackdropExiting,
+    pickerContainerEntering,
+    pickerContainerExiting,
+    onValueChange,
+    open,
+    handleDelete,
+    handleCancel,
+    handleDone,
+  } = useSelectPickerUseCase<ItemT>(props);
+  const {
+    items,
+    selectedItemKey,
+    placeholder,
+    textInputProps,
+    keyExtractor,
+    pickerAccessoryComponent,
+    pickerItemsComponent,
+    textInputComponent,
+    pickerItemsContainerProps: {style: pickerItemsContainerStyle, ...pickerItemsContainerProps} = {},
+    pickerProps,
+    pickerBackdropProps: {entering: backdropEntering, exiting: backdropExiting, ...pickerBackdropProps} = {},
+    pickerContainerProps: {
+      entering: containerEntering,
+      exiting: containerExiting,
+      style: pickerContainerStyle,
+      ...pickerContainerProps
+    } = {},
+    pickerAccessoryProps,
+  } = props;
+
+  return (
+    <>
+      <PickerBackdrop
+        isVisible={isVisible}
+        onPress={handleBackdropPress}
+        entering={pickerBackdropEntering}
+        exiting={pickerBackdropExiting}
+        {...pickerBackdropProps}>
+        <PickerContainer
+          isVisible={isVisible}
+          entering={pickerContainerEntering}
+          exiting={pickerContainerExiting}
+          style={StyleSheet.flatten([styles.pickerContainer, pickerContainerStyle])}
+          {...pickerContainerProps}>
+          {pickerAccessoryComponent ? (
+            pickerAccessoryComponent
+          ) : (
+            <DefaultPickerAccessory
+              onDelete={handleDelete}
+              onCancel={handleCancel}
+              onDone={handleDone}
+              {...pickerAccessoryProps}
+            />
+          )}
+          <View {...pickerItemsContainerProps}>
+            {pickerItemsComponent ? (
+              pickerItemsComponent
+            ) : (
+              <SelectPickerItems
+                selectedValue={selectedItemKey}
+                items={items}
+                onValueChange={onValueChange}
+                keyExtractor={keyExtractor}
+                {...pickerProps}
+              />
+            )}
+          </View>
+        </PickerContainer>
+      </PickerBackdrop>
+      <Pressable onPress={open} testID="pressableContainer">
+        <View pointerEvents="box-only">
+          {textInputComponent ? (
+            textInputComponent
+          ) : (
+            // テキスト入力とスタイルを合わせるために、TextではなくTextInputを使用する
+            <TextInput placeholder={placeholder} value={inputValue} editable={false} {...textInputProps} />
+          )}
+        </View>
+      </Pressable>
+    </>
+  );
 };
+
+const styles = StyleSheet.create({
+  pickerContainer: {
+    backgroundColor: 'white',
+  },
+});
