--- example-app/SantokuApp/src/components/picker/useSelectPickerIOSUseCase.ts
@@ -1,7 +1,0 @@
-import {SelectPickerProps} from './SelectPicker';
-import {useSelectPickerUseCase} from './useSelectPickerUseCase';
-
-export const useSelectPickerIOSUseCase = <ItemT extends unknown>(props: SelectPickerProps<ItemT>) => {
-  const {getSelectedItem, ...selectPickerUseCase} = useSelectPickerUseCase<ItemT>(props);
-  return {...selectPickerUseCase};
-};
